refactor: 一覧画面で使用するファイルを sponsor/list ディレクトリへ移動
